Checking in  v5_3_0. Adding License Deleting obsolete files

+++ cps_pp/include/alg/alg_fourier_prop.h
@@ -34,7 +34,7 @@ private:
   char* cname;
 
   //! Node checkerboard size of the fermion field
-  int f_size;
+  size_t f_size;
 
 protected:
