+++ cps_pp/include/util/qio_readGenericFields.h
@@ -33,7 +33,7 @@ public:
     \param field    pointer to field
    */
   qio_readGenericFields(char *outfile,
-			 const int n_fields, const int f_size_per_site,  void *field,
+			 const int n_fields, const size_t f_size_per_site,  void *field,
 			 int argc, char *argv[], int volFormat=QIO_VOLFMT,
 			 FP_FORMAT floatFormat=FP_AUTOMATIC):
     qio_init(argc, argv),
@@ -52,7 +52,7 @@ public:
     \param traj trajectory number
    */
   qio_readGenericFields(char *outfile,
-			 const int n_fields, const int f_size_per_site,  void *field,
+			 const int n_fields, const size_t f_size_per_site,  void *field,
 			 const char * ensemble_id,
 			 const char * ensemble_label,
 			 const int traj,
@@ -101,7 +101,7 @@ public:
   */  
 
   void read_genericfields(char *infile,
-			    const int n_fields, const int f_size_per_site,  void *field,
+			    const int n_fields, const size_t f_size_per_site,  void *field,
 			    int volFormat=QIO_VOLFMT,
 			    FP_FORMAT floatFormat=FP_AUTOMATIC);
 
@@ -112,7 +112,7 @@ public:
   */
   void read_genericfields_eo(
 			      char *infile,
-			      const int n_fields, int f_size_per_site,  void *field,
+			      const int n_fields, size_t f_size_per_site,  void *field,
 			      int volFormat=QIO_VOLFMT ,
 			      FP_FORMAT floatFormat=FP_AUTOMATIC )
   {
