+++ cps_pp/production_code/alg_lanczos/v2/main.C
@@ -51,7 +51,7 @@ LanczosArg lanczos_arg;
 QPropWArg qp_arg;
 
 
-void movefloattoFloat (Float * out, float *in, int f_size);
+void movefloattoFloat (Float * out, float *in, size_t f_size);
 
 
 static void SetZmobiusPC(int flag)
@@ -176,7 +176,7 @@ void comp_read_eigenvectors (Lattice & lattice, int eig_start=0,int num_eig=0)
       ecache->alloc (evecname_bc, neig, fsize);
       {				//read in only
 	const int n_fields = GJP.SnodeSites ();
-	const int f_size_per_site = lattice.FsiteSize () / n_fields / 2;
+	const size_t f_size_per_site = lattice.FsiteSize () / n_fields / 2;
 	EigenContainer eigcon (lattice, evecname_bc, neig, f_size_per_site / 2,
 			       n_fields, ecache);
 	// factor of 2 for single-prec.
@@ -328,7 +328,7 @@ if (do_meas){
 }
 
 
-void movefloattoFloat (Float * out, float *in, int f_size)
+void movefloattoFloat (Float * out, float *in, size_t f_size)
 {
 
   float flt;
