+++ cps_pp/src/util/dirac_op/d_op_base/qcdoc/minvcg.C
@@ -74,7 +74,7 @@ int DiracOp::MInvCG(Vector **psi_slow, Vector *chi, Float chi_norm, Float *mass,
   //------------------------------------------------------------------
 
   int iz, k, s;
-  int f_size;
+  size_t f_size;
   if(lat.Fclass() == F_CLASS_CLOVER)
     f_size = lat.FsiteSize()*GJP.VolNodeSites() / 2;
   else
