+++ cps_pp/src/util/dirac_op/d_op_dwf/sse/dwf_init_sse.C
@@ -73,7 +73,7 @@ void dwf_init(Dwf *dwf_p)
 //------------------------------------------------------------------
 // Allocate memory for two temporary fermion checkerboard fields  
 //------------------------------------------------------------------
-  int f_size = 24 * GJP.VolNodeSites() * GJP.SnodeSites() / 2; 
+  size_t f_size = 24 * GJP.VolNodeSites() * GJP.SnodeSites() / 2; 
 
 //  dwf_p->frm_tmp1 = (IFloat *) smalloc(f_size*sizeof(IFloat));
   dwf_p->frm_tmp1 = (IFloat *) smalloc(cname,fname, "frm_tmp1", f_size*sizeof(IFloat));
