+++ cps_pp/src/util/dirac_op/d_op_wilson/bg_bagel/d_op_wilson.C
@@ -476,7 +476,7 @@ void DiracOpWilson::CalcHmdForceVecs(Vector *chi)
 
   Vector *chi_new, *rho, *psi, *sigma ;
 
-  int f_size_cb = 12 * GJP.VolNodeSites() ;
+  size_t f_size_cb = 12 * GJP.VolNodeSites() ;
 
   chi_new = f_out ;
