+++ cps_pp/src/util/dirac_op/d_op_zmobius/noarch/zmobius_dminus.C
@@ -62,7 +62,7 @@ void zmobius_dminus(Vector *out,
   // Initializations
   //----------------------------------------------------------------
   ls = mobius_lib_arg->ls;
-  const int f_size = 24 * mobius_lib_arg->vol_4d / 2;
+  const size_t f_size = 24 * mobius_lib_arg->vol_4d / 2;
 
 
   frm_in = (IFloat *) in;
